add bulk resource and location endpoints (fixes #10)
This involves a major reorganization of how downloadable resources
are handled. A lot of code was moved out of models and into the view
and an accessory module. The big change to the protocol is that
the location query doesn't check to see if the resource file actually exists;
users will find this out when they try to download the resource. This
should streamline queries.